Refactor project structure to improve encapsulation and organization
- Moved `src/highlight.rs` to `src/tools/highlight.rs` to unclutter the root directory.
- Created `src/tools/mod.rs` to manage tool modules.
- Updated `src/lib.rs` to export `tools` and re-export `highlight` for backward compatibility.
- Encapsulated `parser::builder` by changing its visibility to `pub(crate)` in `src/parser/mod.rs`, preventing internal implementation details from leaking into the public API.
- Updated `.jules/atlas.md` with a journal entry documenting the structural changes.
