+++ src/lib.rs
@@ -49,9 +49,11 @@ pub mod ast;
 pub mod codegen;
 pub mod errors;
 pub mod grammar;
-pub mod highlight;
 pub mod morphology;
 pub mod parser;
 pub mod report;
 pub mod semantic;
 pub mod text;
+pub mod tools;
+
+pub use tools::highlight;
